Update plot_pie so bbox_props is function param
Would it be possible to add bbox_props as a parameter to the plot_pie function in order to make the chart even more customisable? My issue is that I'm using a dark theme for the Jupyter notebooks and I created a custom matplotlib style by creating a rcParams file. I've been looking on how to change the bbox style from there but without success. Since my text color is white, I need the bbox background to be customisable. Maybe there's a way but I couldn't figure it out. That's why I came here to propose this change.

Probably it could also be applied to some of the other Plot Functions. I'm down to apply the changes if you are interested in it and, also, following the format you require to do so.

Podemos hablar en español si lo deseas también. Un saludo y muchas gracias por la librería.